+++ sql/pg_clickhouse.sql
@@ -28,6 +28,17 @@ RETURNS TEXT
 AS 'MODULE_PATHNAME'
 LANGUAGE C STRICT;
 
+-- Execute query against foreign server, mapping ClickHouse result to Postgres
+-- types named in caller's column definition list, e.g.
+-- SELECT FROM clickhouse_query('srv', 'SELECT a, b FROM t') AS (a int, b text);
+CREATE FUNCTION clickhouse_query(TEXT, TEXT)
+RETURNS SETOF record
+AS 'MODULE_PATHNAME'
+LANGUAGE C STRICT;
+
+-- As with clickhouse_raw_query(), don't let PUBLIC run arbitrary remote queries.
+REVOKE EXECUTE ON FUNCTION clickhouse_query(text, text) FROM PUBLIC;
+
 CREATE FUNCTION clickhouse_fdw_validator(text[], oid)
 RETURNS VOID
 AS 'MODULE_PATHNAME'